+++ src/lib/gadgets/sha256.ts
@@ -5,72 +5,101 @@ import { Provable } from '../provable.js';
 
 export { SHA256 };
 
-// constants §4.2.2
-const K = [
-  0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1,
-  0x923f82a4, 0xab1c5ed5, 0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3,
-  0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174, 0xe49b69c1, 0xefbe4786,
-  0x0fc19dc6, 0x240ca1cc, 0x2de92c6f, 0x4a7484aa, 0x5cb0a9dc, 0x76f988da,
-  0x983e5152, 0xa831c66d, 0xb00327c8, 0xbf597fc7, 0xc6e00bf3, 0xd5a79147,
-  0x06ca6351, 0x14292967, 0x27b70a85, 0x2e1b2138, 0x4d2c6dfc, 0x53380d13,
-  0x650a7354, 0x766a0abb, 0x81c2c92e, 0x92722c85, 0xa2bfe8a1, 0xa81a664b,
-  0xc24b8b70, 0xc76c51a3, 0xd192e819, 0xd6990624, 0xf40e3585, 0x106aa070,
-  0x19a4c116, 0x1e376c08, 0x2748774c, 0x34b0bcb5, 0x391c0cb3, 0x4ed8aa4a,
-  0x5b9cca4f, 0x682e6ff3, 0x748f82ee, 0x78a5636f, 0x84c87814, 0x8cc70208,
-  0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2,
-].map((k) => UInt32.from(k));
+function processStringToMessageBlocks(s: string) {
+  let msgBits = s
+    .split('')
+    .map((c) => {
+      let binary = c.charCodeAt(0).toString(2);
+      return '00000000'.substr(binary.length) + binary;
+    })
+    .join('');
 
-// initial hash values §5.3.3
-const H = [
-  0x6a09e667, 0xbb67ae85, 0x3c6ef372, 0xa54ff53a, 0x510e527f, 0x9b05688c,
-  0x1f83d9ab, 0x5be0cd19,
-].map((h) => UInt32.from(h));
+  let l = msgBits.length;
+  msgBits = msgBits + '1';
+
+  // calculate k in l + 1 +k = 448 mod 512
+  let remainder = (448 - (l + 1)) % 512;
+
+  let k = (remainder + 512) % 512;
+  let padding = '0'.repeat(k);
+  msgBits = msgBits + padding;
+  let lBits = l.toString(2);
+  msgBits = msgBits + '0'.repeat(64 - lBits.length) + lBits;
+
+  let bitBlocks32 = [];
+  for (let i = 0; i < msgBits.length; i += 32) {
+    bitBlocks32.push(UInt32.from(BigInt('0b' + msgBits.substr(i, 32))));
+  }
+
+  let lengthBlocks = bitBlocks32.length;
+  let blocks = [];
+  for (let i = 0; i < lengthBlocks; i += 16) {
+    let block = bitBlocks32.slice(i, i + 16);
+    blocks.push(block);
+  }
+  return blocks;
+}
 
 const SHA256 = {
-  hash(msg: string) {
-    // preprocessing §6.2
+  hash(data: UInt32[][]) {
+    // constants §4.2.2
+    const K = [
+      0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1,
+      0x923f82a4, 0xab1c5ed5, 0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3,
+      0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174, 0xe49b69c1, 0xefbe4786,
+      0x0fc19dc6, 0x240ca1cc, 0x2de92c6f, 0x4a7484aa, 0x5cb0a9dc, 0x76f988da,
+      0x983e5152, 0xa831c66d, 0xb00327c8, 0xbf597fc7, 0xc6e00bf3, 0xd5a79147,
+      0x06ca6351, 0x14292967, 0x27b70a85, 0x2e1b2138, 0x4d2c6dfc, 0x53380d13,
+      0x650a7354, 0x766a0abb, 0x81c2c92e, 0x92722c85, 0xa2bfe8a1, 0xa81a664b,
+      0xc24b8b70, 0xc76c51a3, 0xd192e819, 0xd6990624, 0xf40e3585, 0x106aa070,
+      0x19a4c116, 0x1e376c08, 0x2748774c, 0x34b0bcb5, 0x391c0cb3, 0x4ed8aa4a,
+      0x5b9cca4f, 0x682e6ff3, 0x748f82ee, 0x78a5636f, 0x84c87814, 0x8cc70208,
+      0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2,
+    ].map((k) => UInt32.from(k));
+
+    // initial hash values §5.3.3
+    const H = [
+      0x6a09e667, 0xbb67ae85, 0x3c6ef372, 0xa54ff53a, 0x510e527f, 0x9b05688c,
+      0x1f83d9ab, 0x5be0cd19,
+    ].map((h) => UInt32.from(h));
+
+    // TODO: correct dynamic preprocessing §6.2
     // padding the message $5.1.1 into blocks that are a multiple of 512
-    let l = 8 * 8;
-    let k = 448 - (l + 1);
-    let padding = [1, ...new Array(k).fill(0)];
+    let messageBlocks = data;
 
-    let messageBlocks = [
-      // M1
-      new Array(16).fill(UInt32.from(1)),
-      new Array(16).fill(UInt32.from(1)),
-    ];
+    const N = messageBlocks.length;
 
-    for (let i = 0; i < messageBlocks.length; i++) {
+    for (let i = 0; i < N; i++) {
       const M = messageBlocks[i];
       // for each message block of 16 x 32 bytes do:
       const W = new Array(64);
 
       // prepare message block
-      for (let t = 0; t < 16; t++) W[t] = M[t];
-      for (let t = 16; t < 64; t++) {
+      for (let t = 0; t <= 15; t++) W[t] = M[t];
+      for (let t = 16; t <= 63; t++) {
         W[t] = DeltaOne(W[t - 2])
           .addMod32(W[t - 7])
-          .add(DeltaZero(W[t - 15]).add(W[t - 16]));
+          .addMod32(DeltaZero(W[t - 15]).addMod32(W[t - 16]));
       }
       // initialize working variables
-      let a = H[0],
-        b = H[1],
-        c = H[2],
-        d = H[3],
-        e = H[4],
-        f = H[5],
-        g = H[6],
-        h = H[7];
+      let a = H[0];
+      let b = H[1];
+      let c = H[2];
+      let d = H[3];
+      let e = H[4];
+      let f = H[5];
+      let g = H[6];
+      let h = H[7];
 
       // main loop
-      for (let t = 0; t < 64; t++) {
+      for (let t = 0; t <= 63; t++) {
         const T1 = h
           .addMod32(SigmaOne(e))
           .addMod32(Ch(e, f, g))
           .addMod32(K[t])
           .addMod32(W[t]);
 
-        const T2 = SigmaOne(a).addMod32(Maj(a, b, c));
+        const T2 = SigmaZero(a).addMod32(Maj(a, b, c));
 
         h = g;
         g = f;
@@ -94,20 +123,10 @@ const SHA256 = {
       H[7] = H[7].addMod32(h);
     }
 
-    /*
-    let hex = '';
-    for (let h = 0; h < H.length; h++)
-      hex = hex + ('00000000' + H[h].toBigint().toString(16)).slice(-8);
-    console.log(hex);
-    */
+    return H;
   },
+  processStringToMessageBlocks: processStringToMessageBlocks,
 };
-
-let cs = Provable.constraintSystem(() => SHA256.hash('abc'));
-console.log(cs);
-/* function toUInt8(msg: string) {
-  return msg.split('').map((c) => UInt8.from(c.charCodeAt(0)));
-} */
 
 function Ch(x: UInt32, y: UInt32, z: UInt32) {
   let xAndY = x.and(y);
